requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
langchain>=0.1.0
langgraph>=0.0.10
langchain-openai>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
firebase-admin>=6.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
apscheduler>=3.10.0
python-multipart>=0.0.6
jinja2>=3.1.0
